requirements.txt
python-dotenv
flask
flask_sqlalchemy
flask_login
collections
spacy
nltk
chromadb
langchain
openai